requirements.txt
discord.py
flask
gunicorn
psycopg2-binary
python-dotenv
pytz